requirements_gpu.txt
keras
tensorflow-gpu==1.15.0
pandas
numpy
tqdm
docopt
coloredlogs
jupyterlab=
altair
seaborn
